% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_cadmium.R
\name{analyse_cadmium3}
\alias{analyse_cadmium3}
\title{Third and last step in ICP-MS R analysis for cadmium_111 without adjusting for drift.}
\usage{
analyse_cadmium3(a, b, c, d, e, f, g, h, i, j)
}
\arguments{
\item{a}{Enter Cadmium_analysis}

\item{b}{Name of standards used for monitoring drift e.g. "1 ppb standard"}

\item{c}{Name of first standard curve sample e.g. "0.0004 ppb metal standard"}

\item{d}{Name of second standard curve sample e.g. "0.002 ppb metal standard"}

\item{e}{Name of third standard curve sample e.g. "0.02 ppb metal standard"}

\item{f}{Name of fourth standard curve sample e.g. "0.1 ppb metal standard"}

\item{g}{Name of fifth standard curve sample e.g. "0.5 ppb metal standard"}

\item{h}{Name of sixth standard curve sample e.g. "1 ppb metal standard"}

\item{i}{Name of seventh standard curve sample e.g. "5 ppb metal standard"}

\item{j}{Name of eighth standard curve sample e.g. "10 ppb metal standard"}
}
\value{
Dataframe and .csv file with all cadmium results
}
\description{
Transforms cadmium (Cd_111) cps outputs to parts per billion, accounting for In_115 internal standard,
plotting a graph for baseline drift using 1 ppb standards but not adjusting values according to drift,
using blank values to deduct background signal, plotting a cadmium standard curve,
accounting for 1/50 dilution during sample prep, and summarising cadmium results.
Results output includes a cadmium_results dataframe
and a .csv file in a new 'Results' folder in the R working directory
}
\details{
Enter Cadmium_analysis, then names of standard for drift and standards for standard curve
e.g. analyse_cadmium3(Cadmium_analyis, "1 ppb standard", "0.0004 ppb metal standard", "0.002 ppb metal standard",
"0.02 ppb metal standard", "0.1 ppb metal standard", "0.5 ppb metal standard", "1 ppb metal standard",
"5 ppb metal standard", "10 ppb metal standard")

Using this exact notation ("___ ppb metal standard") for standard curve samples is critical
}
