% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_lead_withdrift.R
\name{analyse_lead_withdrift3}
\alias{analyse_lead_withdrift3}
\title{Third and last step in ICP-MS R analysis for lead_208 when adjusting for drift.}
\usage{
analyse_lead_withdrift3(a, b, c, d, e, f, g, h, i, j)
}
\arguments{
\item{a}{Enter Lead_analysis}

\item{b}{Name of standards used for monitoring drift e.g. "1 ppb standard"}

\item{c}{Name of first standard curve sample e.g. "0.0004 ppb metal standard"}

\item{d}{Name of second standard curve sample e.g. "0.002 ppb metal standard"}

\item{e}{Name of third standard curve sample e.g. "0.02 ppb metal standard"}

\item{f}{Name of fourth standard curve sample e.g. "0.1 ppb metal standard"}

\item{g}{Name of fifth standard curve sample e.g. "0.5 ppb metal standard"}

\item{h}{Name of sixth standard curve sample e.g. "1 ppb metal standard"}

\item{i}{Name of seventh standard curve sample e.g. "5 ppb metal standard"}

\item{j}{Name of eighth standard curve sample e.g. "10 ppb metal standard"}
}
\value{
Dataframe and .csv file with all lead results
}
\description{
Transforms lead (Pb_208) cps outputs to parts per billion, accounting for Ir_193 internal standard,
plotting a graph for baseline drift using 1 ppb standards and adjusting values according to drift,
using blank values to deduct background signal, plotting a lead standard curve,
accounting for 1/50 dilution during sample prep, and summarising lead results.
Results output includes a lead_results dataframe
and a .csv file in a new 'Results' folder in the R working directory
}
\details{
Enter Lead_analysis, then names of standard for drift and standards for standard curve
e.g. analyse_lead_withdrift3(Lead_analyis, "1 ppb standard", "0.0004 ppb metal standard", "0.002 ppb metal standard",
"0.02 ppb metal standard", "0.1 ppb metal standard", "0.5 ppb metal standard", "1 ppb metal standard",
"5 ppb metal standard", "10 ppb metal standard")

Using this exact notation ("___ ppb metal standard") for standard curve samples is critical
}
