% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_all.R
\name{setup_all1}
\alias{setup_all1}
\title{First step in ICP-MS R analysis}
\usage{
setup_all1(x)
}
\arguments{
\item{x}{name of ICP-MS results file}
}
\value{
Loaded packages and a ICP-MS results file made into the "intensity" dataframe
}
\description{
Installs necessary packages and loads ICP-MS result.
Enter name of saved ICP-MS results file e.g. setup_all1("Sample_run_HeavyMetalstest_modified.csv")
}
