% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_cadmium.R
\name{setup_cadmium2}
\alias{setup_cadmium2}
\title{Second step in ICP-MS R analysis for cadmium_111 using indium_115 as internal standard.}
\usage{
setup_cadmium2(a, b)
}
\arguments{
\item{a}{Enter "Cd_111"}

\item{b}{Enter Intensities}
}
\value{
Dataframe with cadmium data including RSD and internal standard
}
\description{
Arranges cadmium (Cd_111) data to prepare for analysis by analyse_cadmium3 or analyse_cadmium_withdrift3
Enter Cd_111 e.g. setup_cadmium2("Cd_111", Intensities)
}
