% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_lead.R
\name{setup_lead2}
\alias{setup_lead2}
\title{Second step in ICP-MS R analysis for lead_208 using iridium_193 as internal standard.}
\usage{
setup_lead2(a, b)
}
\arguments{
\item{a}{Enter "Pb_208"}

\item{b}{Enter Intensities}
}
\value{
Dataframe with lead data including RSD and internal standard
}
\description{
Arranges lead (Pb_208) data to prepare for analysis by analyse_lead3 or analyse_lead_withdrift3
Enter Cd_111 e.g. setup_lead2("Pb_208", Intensities)
}
