% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_mercury.R
\name{setup_mercury2}
\alias{setup_mercury2}
\title{Second step in ICP-MS R analysis for mercury_202 using iridium_193 as internal standard.}
\usage{
setup_mercury2(a, b)
}
\arguments{
\item{a}{Enter "Hg_202"}

\item{b}{Enter Intensities}
}
\value{
Dataframe with mercury data including RSD and internal standard
}
\description{
Arranges mercury (Hg_202) data to prepare for analysis by analyse_mercury3 or analyse_mercury_withdrift3
Enter Hg_202 e.g. setup_mercury2("Hg_202", Intensities)
}
